requirements.txt
openai
nltk
scikit-learn
rouge-score
python-dotenv